Added ASL FileRequest/MessageBox, locale OpenCatalog, and icon DiskObject APIs on the builtin amiga module Added amiga path converters, Dos helpers via _amigados, and OS4-style asl/catalog/icon/arexx shims under Lib/site-python Added AmigaTests coverage for GUI extras and ziplib Fixed site.py so Amiga puts Lib/site-python on sys.path and loads _amigados after bootstrap Fixed pyexpat/expat and related Amiga build wiring for the VBCC port Removed standalone Lib/Dos.py in favour of amiga-module Dos APIs and a legacy dos shim Removed private contact emails and outdated 1.6/2.0-only wording from classic Amiga docs Updated classic documentation for AmigaPython 2.7.18 (Python/docs, Amiga_Misc, README.AMIGA, Embed, CHANGES, installer, RunTest, cheatsheet)

+++ Python/RunTest.py
@@ -1,29 +1,48 @@
 
 def dotest():
-	print 'This script will now start the Python 1.5 test set.'
-	print 'Depending on the speed of your Amiga, this may take several minutes.'
+	print 'This script will now start the Python 2.7 regression test set.'
+	print 'Depending on the speed of your Amiga, this may take a long time.'
 	print
-	print 'The report at the end should say something like:'
-	print '   32 tests OK.'
-	print '   18 tests skipped:  .....'
-	print '(If you don\'t start AmiTCP first, test_socket and test_select will fail.)'
+	print 'Some tests are skipped on the Amiga.  If you do not have'
+	print 'bsdsocket.library available, test_socket and test_select will fail.'
+	print 'Without usergroup.library, crypt/grp/pwd-style tests may fail.'
 	print
-	print 'To be 100% sure, run this script TWICE:'
-	print 'first, answer \'y\' to the question \'delete .pyc files?\', and the'
-	print 'second time, answer \'n\'.'
+	print 'To be thorough, run this script TWICE:'
+	print "first, answer 'y' to delete .pyc files, then again answering 'n'."
 	print
 
-	# Change to the test directory
-	import sys,os
-	os.chdir('Python:Lib/Python1.5/test')
+	# Change to the test directory under the versioned library
+	import sys, os
+	candidates = [
+		'Python:Lib/test',
+		'Python:Lib/python2.7/test',
+		'Python:lib/test',
+		'Python:lib/python2.7/test',
+	]
+	testdir = None
+	for path in candidates:
+		if os.path.isdir(path):
+			testdir = path
+			break
+	if testdir is None:
+		# Fall back relative to this script / sys.prefix
+		here = os.path.dirname(sys.argv[0])
+		for rel in ('Lib/test', 'lib/test', 'Lib/python2.7/test'):
+			path = os.path.join(here, rel)
+			if os.path.isdir(path):
+				testdir = path
+				break
+	if testdir is None:
+		print 'Could not find Lib/test.  Set ASSIGN Python: and retry.'
+		return
+	os.chdir(testdir)
 
-	if 'y'==raw_input('Delete .pyc files first? '):
-		# Remove all .pyc files
+	if 'y' == raw_input('Delete .pyc files first? '):
 		import glob
 		for pyc in glob.glob('*.pyc'):
 			os.unlink(pyc)
 
-	print 'Testing...'
+	print 'Testing in', testdir, '...'
 	import test.autotest
 	print 'Finished.'
